Se crea test de listar carta correcto
